+++ src/imio/helpers/xhtml.py
@@ -468,3 +468,15 @@ def storeImagesLocally(context,
                                        encoding='utf-8',
                                        pretty_print=pretty_print,
                                        method='html') for x in tree.iterchildren()])
+
+
+def object_link(obj, view='view', attribute='Title', content=''):
+    """ Returns an html link for the given object """
+    href = view and "%s/%s" % (obj.absolute_url(), view) or obj.absolute_url()
+    if not content:
+        if not hasattr(obj, attribute):
+            attribute = 'Title'
+        content = getattr(obj, attribute)
+        if callable(content):
+            content = content()
+    return u'<a href="%s">%s</a>' % (href, safe_unicode(content))