+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
   if `True` (default), the new added metadata are updateid on every
   catalogued objects.
   [gbastien]
+- Added function to return html link for an object
+  [sgeulette]
 
 0.21 (2019-08-13)
 -----------------